Refactor createCard function creating moduls
