+++ src/ui/components/cards/card/components/createCardAge.ts
@@ -1,0 +1,11 @@
+import { type Character } from "../../../../../characters/character/types";
+
+const createCardAge = (character: Character): HTMLElement => {
+  const cardAge = document.createElement("span");
+  cardAge.classList.add("card__age");
+  cardAge.textContent = `Age: ${character.age} years`;
+
+  return cardAge;
+};
+
+export default createCardAge;
